refactor: extract shared bulk-select helpers and fix dialog consistency
- Extract resolveLocationAssetIds and resolveLocationKitIds
  into app/modules/location/bulk-select.server.ts to share
  ALL_SELECTED_KEY resolution logic between assets and kits
- Update both route actions to use the shared helpers
- Switch assets bulk dialog from useLoaderData to useParams
  for consistency with the kits dialog
